requirements.txt
cohere==5.11.0
colorama==0.4.6
groq==0.11.0
python-dotenv==1.0.1
toml==0.10.2
setuptools
ruff
black